{source,target}-vocabulary: simplify
In #18 [1], a discussion around rml:access came to the conclusion
that we can remove it in favor of simplifying the design of Source
and Target. Drop this property and adjust all examples in the
specification part. Additional examples will be updated in a later
stage.

Moreover, a solution for paginated sources (#8) [2] is included
by allowing custom ontologies to be created for custom paginated sources
as there is no standard for paginated Web APIs, all Web APIs have their
own approach of paginating their responses. Therefore, custom ontologies
can be leveraged to deal with these cases.

[1] https://github.com/kg-construct/rml-target-source-spec/issues/18
[2] https://github.com/kg-construct/rml-target-source-spec/issues/8